pyproject.toml
[tool.poetry]
name = "writeTool"
version = "0.0.0"
description = ""
authors = ["Tom <tom@tdpain.net>"]
license = "ISC"

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.5"
fire = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
